moved route functions into seperate files
