+++ anchor-api/lib/index.ts
@@ -1,6 +1,8 @@
 import express from 'express';
 import Article from './models/article.model';
-import { getArticles, handleWebScraper, searchArticles } from './controllers/handleWebScraper';
+import { getArticles } from './controllers/getArticles';
+import {handleWebScraper } from './controllers/handleWebScraper';
+import { searchArticles } from './controllers/searchArticles';
 import dotenv from 'dotenv';
 import mongoose from "mongoose";
 import cors from 'cors';
